perf(box2d): int32 tree links make TreeNode one cache line
Narrow TreeNode.Child1/Child2/Parent to upstream's int32 width so the
node is exactly 64 bytes (one cache line; two nodes per Apple M-series
128-byte fetch) instead of straddling two lines. Indices stay Go int in
every signature and local - the narrow width is storage only, converts
losslessly, and no index approaches 2^31, so results are byte-identical
(all goldens pass unchanged at every worker count).

The dynamic tree is the write-and-read-hot structure in proxy-churn
scenes (insert/remove/balance/refit/rebuild/query are all node-bound),
so this single layout change is the largest serial win of the perf
effort. Same-hour benchstat A/B (n=6, Apple M5 Max):

  MixedRain  100 bodies  serial -53%  workers -41%
  MixedRain  500 bodies  serial -30%  workers -13%
  MixedRain 1000 bodies  serial -31%  workers -15%
  MixedRain 5000 bodies  serial -29%  workers -13%
  Jointed                        -5%
  Pyramid (static tree)          unchanged - the expected signature
  Raycast/Overlap 5000-body queries improve ~10-20% as well.

Implemented during the overnight experiment run (E3); verified and
adjudicated against the keep rule (p=0.002 on every improving row, no
regressions).

+++ pkg/box2d/oracle_misc_internal_test.go
@@ -641,7 +641,7 @@ func TestOracleTreeFreeListInvariant(t *testing.T) {
 	}
 
 	freeCount := 0
-	for freeIndex := tree.freeList; freeIndex != NullIndex; freeIndex = tree.nodes[freeIndex].Parent {
+	for freeIndex := tree.freeList; freeIndex != NullIndex; freeIndex = int(tree.nodes[freeIndex].Parent) {
 		freeCount++
 	}
 
@@ -736,7 +736,7 @@ func TestOracleTreeValidateRejectsCorruption(t *testing.T) {
 			corrupt: func(t *testing.T, tree *DynamicTree) {
 				internal := anyInternal(tree)
 				require.NotEqual(t, NullIndex, internal)
-				tree.nodes[internal].Child1 = tree.nodeCapacity + 5
+				tree.nodes[internal].Child1 = int32(tree.nodeCapacity + 5)
 			},
 		},
 		{
@@ -752,7 +752,7 @@ func TestOracleTreeValidateRejectsCorruption(t *testing.T) {
 			corrupt: func(t *testing.T, tree *DynamicTree) {
 				internal := anyInternal(tree)
 				require.NotEqual(t, NullIndex, internal)
-				tree.nodes[tree.nodes[internal].Child1].Parent = tree.root
+				tree.nodes[tree.nodes[internal].Child1].Parent = int32(tree.root)
 			},
 		},
 		{
@@ -760,7 +760,7 @@ func TestOracleTreeValidateRejectsCorruption(t *testing.T) {
 			corrupt: func(t *testing.T, tree *DynamicTree) {
 				internal := anyInternal(tree)
 				require.NotEqual(t, NullIndex, internal)
-				tree.nodes[tree.nodes[internal].Child2].Parent = tree.root
+				tree.nodes[tree.nodes[internal].Child2].Parent = int32(tree.root)
 			},
 		},
 		{
